core/requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
filelock==3.18.0
fsspec==2025.3.2
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
pillow==11.2.1
setuptools==79.0.1
sqlparse==0.5.3
sympy==1.13.3
torch==2.7.0
torchvision==0.22.0
typing_extensions==4.13.2
tzdata==2025.2
